issues solve

+++ lms/app/api/categories/[id]/route.ts
@@ -1,4 +1,5 @@
 import CategoriesModel from "@/app/models/CategoriesModel";
+import { AdminAuth } from "@/lib/adminAuth";
 import { ApiError, ApiSuccess } from "@/lib/api-response";
 import { NextRequest } from "next/server";
 
@@ -9,6 +10,12 @@ export async function DELETE(
   },
 ) {
   try {
+    const auth = await AdminAuth(req);
+    
+    if (!auth.status) {
+      return ApiError(auth.message || "Admin access required", auth.status);
+    }
+
     const { id } = await context.params;
     await CategoriesModel.findByIdAndDelete(id);
     return ApiSuccess("Category deleted succesfully");
